# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Moodify - AI Emotion-Based Music Recommendation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Song {
  id?: number;
  title: string;
  artist: string;
  genre?: string;
  emotion?: string;
  youtubeId: string;
  thumbnail?: string;
  duration?: number;
  isLiked?: boolean;
}

export interface SearchRequest {
  query: string;
  userId?: string;
}

export interface SearchResponse {
  songs: Song[];
  query: string;
}

export interface PlayRequest {
  youtubeId: string;
  title: string;
  artist: string;
  genre?: string;
  emotion?: string;
  thumbnail?: string;
  duration?: number;
  userId?: string;
}

export interface PlayResponse {
  historyId?: number;
  songId?: number;
  message: string;
}

export interface ProgressRequest {
  historyId: number;
  playTime: number;
  duration: number;
  userId?: string;
}

export interface ProgressResponse {
  playPercentage?: number;
  message: string;
}

export interface SkipRequest {
  historyId: number;
  playTime: number;
  duration: number;
  userId?: string;
}

export interface SkipResponse {
  playPercentage?: number;
  message: string;
}

export interface RecommendRequest {
  youtubeId?: string;
  emotion: string;
  genre?: string;
  artist?: string;
  playPercentage?: number;
  userId?: string;
}

export interface RecommendResponse {
  songs: Song[];
  reason: string;
}

export interface HistoryItem {
  id: number;
  song: Song;
  playPercentage: number;
  timestamp: string;
}

export interface HistoryResponse {
  history: HistoryItem[];
}

export interface LikedSongsResponse {
  songs: Song[];
}

export interface ToggleLikeRequest {
  youtubeId: string;
  title: string;
  artist: string;
  genre?: string;
  emotion?: string;
  thumbnail?: string;
  userId?: string;
}

export interface ToggleLikeResponse {
  liked: boolean;
  message: string;
}

export interface CreatePlaylistRequest {
  name: string;
  userId?: string;
}

export interface PlaylistResponse {
  id: number;
  name: string;
  userId?: string;
  createdAt?: string;
  songCount?: number;
}

export interface PlaylistsResponse {
  playlists: PlaylistResponse[];
}

export interface PlaylistDetailResponse {
  playlist: PlaylistResponse;
  songs: Song[];
}

export interface AddToPlaylistRequest {
  playlistId: number;
  youtubeId: string;
  title: string;
  artist: string;
  genre?: string;
  emotion?: string;
  thumbnail?: string;
  userId?: string;
}

export interface AddToPlaylistResponse {
  message: string;
}

export interface RemoveFromPlaylistRequest {
  youtubeId: string;
}

export interface RemoveFromPlaylistResponse {
  message: string;
}

export type GetHistoryParams = {
  userId?: string;
  limit?: number;
};

export type GetLikedSongsParams = {
  userId?: string;
};

export type GetPlaylistsParams = {
  userId?: string;
};